requirements.txt
pyTelegramBotAPI
python-dotenv
gtts
praw
python_weather
pytz
nltk
numpy
torch
flask
flask-cors
waitress